requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.5
opencv-python>=4.5.4
tqdm>=4.62.3
matplotlib>=3.4.3
scikit-learn>=0.24.2
pandas>=1.3.3
Pillow>=8.3.2
scipy>=1.7.1
pyyaml>=5.4.1
tensorboard>=2.6.0
